Fix the regex used to find the `license_id`
Xcode 8 seems to break the regex, because `EA` doesn't come at the start of the line.

For me, it looks like:
```
$ cat /Applications/Xcode_8.app/Contents/Resources/English.lproj/License.rtf | tail -n3

\fs16 \cf0 EA1421\
8/24/16}%
```

So, I changed the regex to look for word boundaries instead.

Here's the testing I did (in irb), where you can see the old regex (the
first `IO.read…`) fails to match, but the new regex succeeds:
```
>> p = "/Applications/Xcode_8.app/Contents/Resources/English.lproj/License.rtf"
=> "/Applications/Xcode_8.app/Contents/Resources/English.lproj/License.rtf"

>> IO.read(p).match(/^EA\d{4}/)
=> nil

>> IO.read(p).match(/\bEA\d{4}\b/)
=> #<MatchData "EA1421">
```