Replace the infamous Spotlight with system_profiler
This adds 2 seconds to each run of `installed` in my
experience. It's not ideal but it's also not awful.
It returns the same results as far as I can tell in my
tests. The regex expression is also tied to how
xcode-install renames the downloaded Xcode.app files
(i.e. `xcode_path`).